+++ tests/test_visitors/test_ast/test_blocks/test_overlap/test_blocks_overlap.py
@@ -1,5 +1,6 @@
 import pytest
 
+from wemake_python_styleguide.compat.constants import PY310
 from wemake_python_styleguide.violations.best_practices import (
     BlockAndLocalOverlapViolation,
 )
@@ -74,6 +75,27 @@ if overlap := other():
     ...
 """
 
+import_and_match_as = """
+import overlap
+
+match ...:
+    case 1 as overlap: ...
+"""
+
+import_and_match_as_implicit = """
+import overlap
+
+match ...:
+    case overlap: ...
+"""
+
+import_and_match_star = """
+import overlap
+
+match ...:
+    case [1, *overlap]: ...
+"""
+
 # Correct:
 
 unused_variables_overlap1 = """
@@ -121,6 +143,27 @@ def context():
     loop_and_loop1,
     loop_and_loop2,
     import_and_walrus,
+    pytest.param(
+        import_and_match_as,
+        marks=pytest.mark.skipif(
+            not PY310,
+            reason='Pattern matching was added in Python 3.10',
+        )
+    ),
+    pytest.param(
+        import_and_match_as_implicit,
+        marks=pytest.mark.skipif(
+            not PY310,
+            reason='Pattern matching was added in Python 3.10',
+        )
+    ),
+    pytest.param(
+        import_and_match_star,
+        marks=pytest.mark.skipif(
+            not PY310,
+            reason='Pattern matching was added in Python 3.10',
+        )
+    ),
 ])
 def test_block_overlap(
     assert_errors,